Change GenerateValues data to type double
git-svn-id: svn://localhost/vantis/repgen/trunk@116 f059b2e8-4df0-46e1-8ac8-86f86453394b

+++ vsprojects/RSMTenon.Graphing/Graph.cs
@@ -133,7 +133,7 @@ namespace RSMTenon.Graphing
             return numericPoint1;
         }
 
-        protected virtual Values GenerateValues(string formatCode, float[] data)
+        protected virtual Values GenerateValues(string formatCode, double[] data)
         {
             Values values1 = new Values();
